+++ pkgs/dart_mcp/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 0.2.1-wip
 
+- Fix the `protocolLogSink` support when using `MCPClient.connectStdioServer`.
 - Update workflow example to show thinking spinner and input and output token
   usage.
+- Update file system example to support relative paths.
+- Fix a bug in notification handling where leaving off the parameters caused a
+  type exception.
 
 ## 0.2.0